feat: update branch_protection for provider v6.x
We add three new attrs:

1. force_push_bypassers
2. lock_branch
3. restrict_pushes

We only support the one-uple pattern for restrict_pushes, similar to
the previous blocks like required_status_checks.

In service of backwards compatibility we do check to see if a live
branch_protections_v4.push_restrictions array exists and use that as a
fallback, as this was the previous attr used for the same function in
integrations/github v4.

References: https://registry.terraform.io/providers/integrations/github/6.2.1/docs/resources/branch_protection